Compute md5 hash of the file contents
Earlier we were computing the md5 hash of the filename and not the
file contents. This was done for testing purpose. It was also handy
for quick testing as the program didn't need to read each file.

In this commit, the code is modified to compute the md5 hash of the
contents of the file which is what we want to do.