Move code into a separate module for generating snapshot
This commit implements a proper module with specific structs for what
we're calling as a 'Snapshot'. As the name suggests, it's the current
state of all the files under the root dir. The module also provides
functions for 'snapshotting' a directory and printing it's text
representation to stdout.

This commit introduces a breaking change in the text representation of
the snapshot. The 'F' prefix in the earlier version has been replaced
with full words such as 'keep', 'symlink', 'delete'. This is similar
to the options available in git's interactive rebase interface.