Refactor Snapshot struct to remove the 'render_text'
Reason: This function was introducing a possible cyclic
dependency (something that rust/cargo might have allowed, but causes
enough confusion). This became clear at the time of implementing a
similar 'Snapshot::from_text' function which would call an underlying
'textformat::parse' function. But 'render_text' method already
estabilishes a dependence of 'snapshot' module on
'snapshot::textformat' module. So the new fn would cause a kind of
cylclic dependency.

Instead it's better to have a one-way dependency of the sub-module
'snapshot::textformat' depend on the parent module 'snapshot'. The
calling code can be implemented in two steps as follows:

1. For writing snapshot to text - The calling code can construct a
   Snapshot instance and pass it to 'snapshot::textformat::render'.

2. For reding snapshot from text - The calling code can call
   'snapshot::textformat::parse' fn to get a Snapshot instance.

This also aligns with the interface that serde libraries
provide. Hence it's be easier to extend it for other formats besides
text too.

Finally, another thing that falls in line due to this change is that
the call to 'println!' to output the text is pushed further outside
into the main module. This way it'd be easier to implement writing to
a file instead of stdout (controlled by a CLI).