Implement support for symlinks found under the root dir
Symlinks to targets which are also located under the root dir are now
considered when generating the Snapshot. This is actually a critical
part of the functionality because the intention is to allow a user to
reuse an existing (compatible) text representation of a snapshot
multiple times. But it was left out earlier in order to get the basic
functionality working first.