Assert Error in tests using match instead of equality
This is done because it's not recommended to implement the equality
traits 'Eq' (and 'PartialEq' by effect) for enums that represent
errors. Usually errors shouldn't have to be compared. However, there's
a legit use case to assert for specific errors in unit tests. So we're
using 'match' with boolean assertions as a workaround.

Some discussion about this recommendation can be found in this thread
-
https://users.rust-lang.org/t/help-understanding-io-error-and-lack-of-partialeq/13212