Add code for validation of existing snapshot
Note that the validation code added in this commit is not being used
yet and hence marked as 'dead_code'.

This commit also introduces a new struct 'Action' to represent an
action that needs to be performed on the files. The return value of
the 'validate' function returns a vector of actions. As the validation
process requires the all files contained in the snapshot to be scanned
anyway, the same function also constructs the actions to be
taken. This will also be useful for implementing a 'dry-run' mode
easily without much duplication of code and logic.

Tests need to be written for this code.