Optimize code for finding duplicates by comparing file sizes
This optimization is based on the assumption that files with the same
md5 hash will have the same size as well. It's much cheaper to obtain
the size of the contents in files by simply looking at the path
metadata. Only those files that are of sizes same as some other file
are considered as candidates for computing the md5 hash. This
significantly reduces the amount of data that needs to be read.

This commit also refactors the code a bit. As we need the logic for
checking files for validity during both obtaining file sizes as well
computing the hashes, it's now extracted out of the 'try_md5_hash'
function into a new function 'is_path_valid', which is used as a
predicate in the call to 'filter'. This also makes the 'try_md5_hash'
function redundant.