Introduce syntax for specifying symlink source path
In the implementation so far, the symlink source paths were not taken
into account. This is a blocker for implementing the 'apply'
subcommand.

One workaround is to consider one of the paths marked as 'keep' in the
group as the symlink target. But that leads us to the question - what
if there are multiple paths in a group marked as 'keep' and one path
marked as 'symlink'. Which path should the symlink point to.

This commit addresses this problem by extending the line syntax such
that the symlink source path could be specified after a '->'
delimiter. Eg.

  symlink /foo/1.txt -> /bar/1.txt

When dumping the snapshot in text format, the dupenukem tool should
always include the source path. However, this will be implemented in
later commits.

When parsing and validating a snapshot, specifying the symlink source
path is optional, which means if it's not specified, we'd fallback to
the above mentioned workaround. Again, this will be implemented in
later commits.

Tests have been added/modified.