Add an optional source field in FileOp::Symlink variant of the enum
This field indicates the source path for the symlink that is found (at
the time of generating the snapshot) or to be created (at the time of
applying changes for a snapshot).

It's an optional field because the user may or may not specify it in
the snapshot. If it's not specified, one of the paths marked as 'keep'
in the group (of duplicates) will be considered as the default. This
fallback mechanism will however be implemented in a later commit.