Normalize the rootdir to absolute path when generating snapshot
This ensures that the behaviour is consistent regardless of whether
the user specifies relative or absolute path when running the 'find'
command.

The downside is that it makes the snapshot output verbose. This can
however be solved at the time of rendering the text representation of
the snapshot by introducing CLI options to allow the users to choose
whether the snapshot should contain relative or absolute paths. This
will be implemented in a separate commit.