Add code to delete files after taking a backup (optionally)
The code is added to 'fileutil' module along with the tests. It's not
yet used (besides tests), which will be done in subsequent commits.

The tests need shared environment (fixtures) hence need to be run
serially. As 'cargo' runs tests in parallel by default, 'serial_test'
is introduced as a new dev-dependency using which it's possible to
annotate certain tests to be run serially.

Fixtures are implemented using private functions inside the
'fileutil::tests' module for now. Perhaps these can be pulled out into
a separate testutil module that can be shared by other tests as
well. This will be taken up in later commits (Not sure how to achieve
this).