Do not canonicalize the symlink source path
The symlink source path, if specified by the user could be either
relative or absolute. Earlier we were canonicalizing this path which
caused it to always be converted to absolute path during both,
rendering as well as parsing the snapshot.

This is not desirable. It should be upto the user to specify whether
the symlink should be relative or absolute and it can be directly
derived from the path that they specify (if the user doesn't
explicitly specify the symlink source path, we consider relative path
by default).

To achieve this, we're additionally capturing whether the symlink
source path provided by the user is relative or absolute in the
'Action::Symlink' variant at the time of snapshot validation. Then in
the execute step it's ensured that the correct type of symlink is
created.