Always use relative paths in snapshot output
This is a pretty significant change. The snapshot output will now
contain paths relative to the rootdir. This makes the snapshot output
less verbose.

This also means that when reading and parsing the snapshot as input,
these paths have to be converted back to absolute paths. It's easy to
think of it as follows,

 - Internally, the code deals with absolute paths
 - Conversion from absolute paths to relative path and vice versa
   happens at the output boundary and input boundary respectively

Note that this is only applicable to the main paths. The symlink
source paths are always relative to the respective main path.