Fix symlink creation when the source is not specified by the user
The code was broken for this particular case. When user doesn't
specify the symlink source path, it's taken from one of the 'keep'
marked paths of the duplicate group. This is always an absolute
path whereas in this case, a relative symlink is created. The bug was
that the relative source path was computed in relation to the rootdir
and not the target path.

To fix this:

  1. The 'pathdiff' crate has been added as a new dependency.

  2. The 'Action::Symlink' variant of the 'Action' enum has been
     modified to have 'is_explicit' field instead of
     'is_relative'. This is because the difference in behaviour
     depends on 'is_explicit' i.e. whether or not the source is
     explicitly specified by the user. Whereas the 'is_relative' field
     is redundant because it can always be computed from the 'source'
     by calling 'source.is_relative()'.