Handle the case of already deleted file better
When a file is marked with the "delete" action marker and if it's
already doesn't exist (perhaps deleted in a previous 'apply' run),
it's now skipped and the snapshot is considered valid.

Earlier, it caused validation to fail, which would break the
assumption that the 'validate' and 'apply' commands are
non-idempotent.

To fix this, the code is refactored such that verification of actual
checksum of a file against the expected checksum of the group is done
only when it's possible. For e.g. it's skipped for non-existing files
which are marked for deletion.