Compare actual and intended symlink source paths correctly
Earlier the 'validate_path_to_symlink' function was comparing the
actual and intended symlink source paths using equality.

This was causing the validation to unnecessarily fail in case the
actual source path is relative and intended source path is not
explicitly specified by the user. Implicit source path is always
absolute and hence equality of absolute and relative paths would
return false even if they were the same paths.

This case is rather common because when the user incrementally runs
apply using the same snapshot file, the expectation is that validation
will pass even if a symlink is created in previous runs without
explicitly specifying the source path. Thus this bug violated the
promise of idempotent behaviour of the 'apply' command.