Use a sorted vector of FilePaths to find the 'keeper'
We use the term 'keeper' to indicate one of the filepaths in a
duplicate group that has the 'keep' action marker. Earlier it was
computed by looking at the first item in the (unsorted) vector with
FileOp == Keep.

This is a bit problematic because there's a possiblity that at the
time of validating an existing snapshot, the order of items in the
vector may change. Suppose one of the duplicate paths in the same
group is marked as a 'symlink' without explicitly mentioning the
symlink source path, and if the symlink has already been
created (note that the 'apply' command is idempotent which means the
same snapshot file can be repeatedly edited and 'applied'), then there
could be a mismatch between the actual symlink source path and the
intended symlink source path, thereby causing a false positive
validation error.

This commit fixes it by first sorting the FilePaths and taking the
first item with FileOp == Keep. As long as no other duplicate file has
been added for the same group, the implicitly selected symlink source
path (which would also be the actual source path at this point), will
always be obtained as the 'keeper'. Hence this commit only reduces
the chances that a false positive validation error and doesn't
eliminate it completely.