Log the max space that can be freed by deduplication
This is implemented for the find command. Before generating the
snapshot output, it now logs the max space that can be freed by
deduplication i.e. if the user were to remove all the duplicate files,
how much space could be saved at max.

This commit introduces the 'size' crate as a new dependency to
format the total file size in logs.