Refactor to use &Path instead of &PathBuf in some cases
'PathBuf' is an owned type whereas 'Path' holds a reference to
'PathBuf'. Hence it makes sense to accept '&Path' instead of
'&PathBuf' as arguments of functions that only need a shared
reference.

This is a partial change i.e. not all occurrences are changed in this
commit. Remaining will be changed in later commits.